Extract Victory Promise display components
