+++ src/components/Profile/components/index.js
@@ -5,4 +5,4 @@ export { default as OfficeInfoList } from './OfficeInfoList';
 export { default as PersonalInfoList } from './PersonalInfoList';
 export { default as ProfileInfoListItem } from './ProfileInfoListItem';
 export { default as SchedulePanel } from './SchedulePanel';
-export { default as VictoryPromiseDisplay } from './VictoryPromiseDisplay';
+export { default as VictoryPromise } from './VictoryPromise';
